% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/apply_kpi_formatting.R
\name{apply_kpi_formatting}
\alias{apply_kpi_formatting}
\title{Apply formatting to a KPI flextable, using functions from the scales package}
\usage{
apply_kpi_formatting(flextable, actuals_cols = NULL, delta_cols = NULL,
  comma_rows = NULL, percent_rows = NULL, dollar_rows = NULL,
  cents_rows = NULL, decimal_rows = NULL)
}
\arguments{
\item{flextable}{an object from ReporteRs::FlexTable()}

\item{actuals_cols}{vector of indices for columns showing actual performance (not deltas)}

\item{delta_cols}{vector of indices for columns showing pct deltas (all percents)}

\item{comma_rows}{vector of row indices that should apply scales::comma()}

\item{percent_rows}{vector of row indices that should apply scales::percent()}

\item{dollar_rows}{vector of row indices that should apply scales::dollar()}

\item{cents_rows}{vector of row indices that should apply scales::dollar() with cents}

\item{decimal_rows}{vector of row indices that should have comma-delimited thousands and also 2 decimal places}
}
\value{
flextable
}
\description{
Apply formatting to a KPI flextable, using functions from the scales package
}

