% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clean_names.R
\name{cleanNames}
\alias{cleanNames}
\title{Use only Lowercase and Underscore in Object Names}
\usage{
cleanNames(x)
}
\arguments{
\item{x}{an object for which a names attribute will be meaningful}
}
\value{
The object \code{x} with names being only lowercase and underscore
}
\description{
Use only Lowercase and Underscore in Object Names
}
\examples{
cleanNames(iris)
}
\seealso{
setNames
}

