% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/contribution_trends.R
\name{contribution_trends}
\alias{contribution_trends}
\title{Plot contribution trends}
\usage{
contribution_trends(df = display_revised, time_col = "week",
  campaign_col = "campaign", grouping = "site_dcm",
  campaign_string = "continuity", agg_fun = summarise_dfa,
  kpis = c("cost", "clicks", "revenue_clickthrough", "revenue_viewthrough"),
  final_week_start = report_dates$final_week_start, nweeks = 7)
}
\arguments{
\item{df}{input data frame}

\item{time_col}{string giving colname with time values}

\item{campaign_col}{string giving campaign colname}

\item{grouping}{string giving field to group_by}

\item{campaign_string}{string to parse for identifying campaign}

\item{agg_fun}{aggregation function, typically a dplyr::summarise() function defined elsewhere}

\item{kpis}{vector of strings giving KPI field names}

\item{final_week_start}{date vector of length 1}

\item{nweeks}{number of weeks to include}
}
\value{
plot
}
\description{
Plot contribution trends
}

