% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/date_diff.R
\name{date_diff}
\alias{date_diff}
\title{Show period-over-period reporting}
\usage{
date_diff(df, date_col, key_colname = "metric", value_colname = "value",
  cols_to_keep = NULL)
}
\arguments{
\item{df}{input data frame}

\item{date_col}{string giving date field colname}

\item{key_colname}{string giving the "key" name (see \code{\link[tidyr]{gather_}})}

\item{value_colname}{string giving the "value" name (see \code{\link[tidyr]{gather_}})}

\item{cols_to_keep}{string vector indicating other columns to keep}
}
\value{
df period-over-period and other breakdowns
}
\description{
Show period-over-period reporting
}

