% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/facetAdjust.R
\name{facetAdjust}
\alias{facetAdjust}
\title{Print a faceted ggplot object with x axis labels repeated on facets}
\usage{
facetAdjust(x, pos = c("up", "down"), newpage = is.null(vp), vp = NULL)
}
\arguments{
\item{x}{a ggplot2 plot object}

\item{pos}{Position of axis labels in faceted plot ("up" or "down")}

\item{newpage}{}

\item{vp}{}
}
\value{
a ggplot object
}
\description{
Print a faceted ggplot object with x axis labels repeated on facets
}
\examples{
facetAdjust(plt, "up")
}

