% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fb_summarise.R
\name{fb_summarise}
\alias{fb_summarise}
\title{Calculate aggregate sums and ratios of facebook metrics}
\usage{
fb_summarise(df)
}
\arguments{
\item{df}{input data frame}
}
\value{
summarized data frame
}
\description{
Calculate aggregate sums and ratios of facebook metrics
}

