% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filter_by_pattern.R
\name{filter_by_pattern}
\alias{filter_by_pattern}
\title{Filter a data frame by values matching a pattern}
\usage{
filter_by_pattern(df, colname, pattern, exclude = FALSE, ignore.case = TRUE)
}
\arguments{
\item{df}{A data frame}

\item{colname}{String column name on which to filter}

\item{pattern}{Regular expression pattern to match for filtering}

\item{exclude}{Exclude cases matching pattern? Defaults to FALSE.}

\item{ignore.case}{Boolean whether string searches should be case-insensitive}
}
\value{
a filtered data frame
}
\description{
Filter a data frame by values matching a pattern
}
\examples{
filter_by_pattern(iris, "Species", "v.r")
}

