% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/find_display_filenames.R
\name{find_display_filenames}
\alias{find_display_filenames}
\title{Grep a chosen folder for certain types of display files}
\usage{
find_display_filenames(folder, type)
}
\arguments{
\item{folder}{string specifying the location of display files}

\item{type}{one of c("adwords", "front_end", or "floodlight")}
}
\value{
string showing full path to the chosen file
}
\description{
Grep a chosen folder for certain types of display files
}

