% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fiscal_calendar}
\alias{fiscal_calendar}
\title{Table of fiscal calendar lookups from 2012-12-01 to 2016-12-31
A dataset containing attributes of dates from 2012-12-01 to 2016-12-31}
\format{A data frame with 1820 rows and 19 variables:
\itemize{
  \item fiscal_year.  character for fiscal year beginning in February
  \item fiscal_month.  character for 3-letter month abbreviation
  \item fiscal_month_num.  numeral character for fiscal month, February is 1
  \item current_date.  Date format, primary key
  \item day_num.  character for day number of fiscal calendar beginning in February
  \item current_week_num.  character for fiscal week numeral beginning in February
  \item current_start_of_week.  date for the Sunday starting the current_date
  \item previous_year_start_of_week.  date for the Sunday starting the current_date - 364
  \item fiscal_quarter.  character for the fiscal quarter 1 through 4, beginning February
  \item previous_year_date.  current_date - 364
  \item day_of_week_num.  character of numerals 1 through 7, Sunday is 1
  \item day_of_week.  character of 3-letter day abbreviation for current_date
  \item season.  chr "Spring" (February through July) or "Fall" (August through January)
  \item previous_week_date.  current_date - 7
  \item total_weeks_in_fiscal_month.  int numer of weeks in fiscal month
  \item total_days_in_fiscal_month.  int number of days in fiscal month
  \item fiscal_month_start.  date of the Sunday beginning the fiscal month
  \item day_num_in_fiscal_month.  character indicating the day number in current fiscal month
  \item calendar_month_start.  character indicating the first date of the calendar month for current_date
}}
\usage{
fiscal_calendar
}
\description{
Table of fiscal calendar lookups from 2012-12-01 to 2016-12-31
A dataset containing attributes of dates from 2012-12-01 to 2016-12-31
}
\keyword{datasets}

