% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/format_deltas.R
\name{format_deltas}
\alias{format_deltas}
\title{Conditional formatting for a time-comparison FlexTable}
\usage{
format_deltas(flextable, data_df, delta_col_nums, cost_ratio_row_nums,
  small_delta = 0.1, big_delta = 0.3)
}
\arguments{
\item{flextable}{An object from ReporteRs::FlexTable()}

\item{data_df}{Data source for the flextable object}

\item{delta_col_nums}{vector giving indices of columns that show percentage comparison}

\item{cost_ratio_row_nums}{vector giving indices of rows that show cost_per_* metrics}

\item{small_delta}{criterion for highlighting minor relative increases or decreases in decimal form.  Default is 0.1}

\item{big_delta}{criterion for highlighting major relative increases or decreases in decimal form.  Default is 0.3}
}
\value{
flextable object with formatting applied
}
\description{
Conditional formatting for a time-comparison FlexTable
}

