% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_metric_position_indices.R
\name{get_metric_position_indices}
\alias{get_metric_position_indices}
\title{Grep a vector of KPI names matching a convention for formatting}
\usage{
get_metric_position_indices(metrics_vec, type = "comma")
}
\arguments{
\item{metrics_vec}{vector of KPIs, likely colnames from a KPI data frame}

\item{type}{choose among c("percent", "dollar", "cost_per", "cents", "comma")}
}
\value{
vector
}
\description{
Grep a vector of KPI names matching a convention for formatting
}

