% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_top_campaign_names.R
\name{get_top_campaign_names}
\alias{get_top_campaign_names}
\title{Get a vector of top campaign names in a sorted dataframe}
\usage{
get_top_campaign_names(df, week_num, n = 3, sort_on = "cost",
  week_lookup_colname = "week", aggregate_function)
}
\arguments{
\item{df}{data frame}

\item{week_num}{fiscal week number for filtering}

\item{n}{number of top campaigns to select (default is 3)}

\item{sort_on}{column to sort descending}

\item{week_lookup_colname}{column name indicating "week"}

\item{aggregate_function}{summarising function}
}
\value{
data frame
}
\description{
Get a vector of top campaign names in a sorted dataframe
}

