% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_week_comparison_lookups.R
\name{get_week_comparison_lookups}
\alias{get_week_comparison_lookups}
\title{Filter a lookup table to relevant dates for WoW and YoY comparisons}
\usage{
get_week_comparison_lookups(fiscal_calendar, final_week_num, this_fiscal_year)
}
\arguments{
\item{fiscal_calendar}{dataframe lookup table}

\item{final_week_num}{fiscal week number as character}

\item{this_fiscal_year}{fiscal year as character}
}
\value{
filtered data frame with date information
}
\description{
Filter a lookup table to relevant dates for WoW and YoY comparisons
}

