% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/join_dcm_frames.R
\name{join_dcm_frames}
\alias{join_dcm_frames}
\title{Merge floodlight and basic-report data frames}
\usage{
join_dcm_frames(front_end_df, floodlight_df)
}
\arguments{
\item{front_end_df}{dataframe with "Basic" DoubleClick report data}

\item{floodlight_df}{dataframe with Basic Floodlight report data}
}
\value{
merged data frame
}
\description{
Merge floodlight and basic-report data frames
}

