% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/list_fiscal_details.R
\name{list_fiscal_details}
\alias{list_fiscal_details}
\title{Produce a list of common fiscal information, using a lookup table}
\usage{
list_fiscal_details(fiscal_calendar,
  week_num_lookup_colname = "current_week_num",
  date_lookup_colname = "current_date",
  fiscal_year_lookup_colname = "fiscal_year", current_date = NULL)
}
\arguments{
\item{fiscal_calendar}{data frame with lookups.  An example is in \code{data(fiscal_calendar)}}

\item{week_num_lookup_colname}{column name for fiscal week number in fiscal_calendar}

\item{date_lookup_colname}{column name indicating date in fiscal_calendar}

\item{fiscal_year_lookup_colname}{column name indicating fiscal year in fiscal_calendar}

\item{current_date}{defaults to \code{\link[lubridate]{today}}.}
}
\value{
list of common fiscal date info
}
\description{
Produce a list of common fiscal information, using a lookup table
}
\examples{
list_fiscal_details(fiscal_calendar)
}

