% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load_kenshoo_ftp.R
\name{load_kenshoo_ftp}
\alias{load_kenshoo_ftp}
\title{Load a Kenshoo report from an FTP location}
\usage{
load_kenshoo_ftp(report_name, ftp_address = "ftp.kenshoo.com",
  is.csv = TRUE, username = NULL, password = NULL, save.creds = FALSE,
  creds.file = NULL)
}
\arguments{
\item{report_name}{String of the report filename.  ".csv" is appended if you exclude it.}

\item{ftp_address}{String with the domain name.  Default is "ftp.kenshoo.com"}

\item{is.csv}{Boolean if the file is csv.}

\item{username}{FTP site username.}

\item{password}{FTP site password. Can be left as NULL if this function is run in the current R session or if the Kenshoo.FTP.Creds list is saved on disk}

\item{save.creds}{Boolean to save Kenshoo.FTP.Creds to disk for next time.}

\item{creds.file}{String to supply location of already-saved Kenshoo.FTP.Creds. If NULL, the function will check "~/Kenshoo.FTP.Creds"}
}
\value{
A data frame for the specified file.
}
\description{
Load a Kenshoo CSV report from an FTP location
}

