% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/proper.R
\name{proper}
\alias{proper}
\title{Convert a string to proper case}
\usage{
proper(str)
}
\arguments{
\item{str}{a character string or vector of strings}
}
\value{
string in proper case
}
\description{
Convert a string to proper case
}
\examples{
proper("abc")
}

