% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/proper_col.R
\name{proper_col}
\alias{proper_col}
\title{Make a "metric" column proper}
\usage{
proper_col(df, colname, levels_vec = NULL)
}
\arguments{
\item{df}{a dataframe with a column named "metric"}

\item{colname}{String giving the name of the column whose values should be made characters in proper case.}

\item{levels_vec}{an optional vector of strings giving levels of a factor.}
}
\value{
data frame with "metric" column in proper case
}
\description{
Make a "metric" column proper
}
\examples{
proper_col(CO2, "Treatment")
}

