% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/proper_metric_col.R
\name{proper_metric_col}
\alias{proper_metric_col}
\title{Make a "metric" column proper}
\usage{
proper_metric_col(df, metric_levels_vec = NULL)
}
\arguments{
\item{df}{a dataframe with a column named "metric"}

\item{metric_levels_vec}{optional character vector indicating the order of metrics as factor.  This must include ALL unique levels in the metric column to avoid NAs}
}
\value{
data frame with "metric" column in proper case
}
\description{
Make a "metric" column proper
}
\examples{
iris \%>\% tidyr::gather(metric, value, -Species) \%>\% proper_metric_col
}

