% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/proper_names.R
\name{proper_names}
\alias{proper_names}
\title{Convert column names to proper case and convert underscore to spaces}
\usage{
proper_names(df, underscore_to_spaces = TRUE)
}
\arguments{
\item{df}{a dataframe with column names}

\item{underscore_to_spaces}{boolean on whether to convert underscore to space in column names}
}
\value{
dataframe with adjusted column names
}
\description{
Convert column names to proper case and convert underscore to spaces
}
\examples{
proper_names(mtcars)
}

