% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_dcm_file.R
\name{read_dcm_file}
\alias{read_dcm_file}
\title{Read a DoubleClick file accounting for extra header rows and totals column}
\usage{
read_dcm_file(filename, skip = 0)
}
\arguments{
\item{filename}{a character string giving the path to a DoubleClick report CSV}

\item{skip}{number of rows to skip if known.  If set at 0, will search for "Report Fields" within first 100 rows.}
}
\value{
a data frame
}
\description{
Read a DoubleClick file accounting for extra header rows and totals column
}

