% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_query.R
\name{read_query}
\alias{read_query}
\title{Read in text from a SQL text file}
\usage{
read_query(query_filepath)
}
\arguments{
\item{query_filepath}{a full path to a SQL text file}
}
\value{
a string with clean query text
}
\description{
Read in text from a SQL text file
}

