% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/return_match.R
\name{return_match}
\alias{return_match}
\title{Return matched regexp group}
\usage{
return_match(string, pattern)
}
\arguments{
\item{string}{a string or vector of strings}

\item{pattern}{a regexp string including parens to find a group}
}
\value{
a vector isolating the matching group, or NA where match not found
}
\description{
Return matched regexp group
}
\examples{
return_match(row.names(mtcars), "(^[^\\\\s]+)")
}

