% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/segment_trend_plot.R
\name{segment_trend_plot}
\alias{segment_trend_plot}
\title{Trended geom_barplot with segmented trended geom_barplots on top}
\usage{
segment_trend_plot(df, trend_col, segmentation_col)
}
\arguments{
\item{df}{input data frame}

\item{trend_col}{column with top-level trends}

\item{segmentation_col}{column for lower_level segmentation}
}
\value{
printed plot
}
\description{
Trended geom_barplot with segmented trended geom_barplots on top
}

