% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/social_subset.R
\name{social_subset}
\alias{social_subset}
\title{Filter/arrange/aggregate/sort Social Data}
\usage{
social_subset(df, post_type = c("engagement", "offer", "clicks", "road_south",
  "biltmore", "fanning"), date_range = c("current_month_to_date",
  "last_month", "last_two_months", "fytd", "fytd_last_month", "trended_months"),
  date_field, return_type = c("table", "chart"), parsing_config_list,
  date_config_list, field_config_list, agg_fun, group_vec = NULL)
}
\arguments{
\item{df}{A raw data frame.}

\item{post_type}{String indicating post grouping.  Must match a name in the field_config and pattern_config lists.}

\item{date_range}{String indicating date range from which to select cases.  Must match a name in the supplied date_config list.}

\item{date_field}{A string giving the field name that identifies cases by date.}

\item{return_type}{A string to specify "table" or "chart" for use in KPI field selection with field_config_list.}

\item{date_config_list}{A list that names several seq.Date's, e.g. last_month}

\item{field_config_list}{A list that specifies KPI field names for each post_type, for use in charting, sort order, and table display.}

\item{agg_fun}{A HARDCODED dplyr::summarise() function providing all aggregates for a particular data set.}

\item{group_vec}{A string vector providing categorical field names by which to group the final data set.}
}
\value{
A subsetted and filtered data frame
}
\description{
Filter/arrange/aggregate/sort Social Data
}
\examples{
social_subset(fb_df, post_type = "clicks", date_range = "fytd", date_field = "start_date", return_type = "table", parsing_config_list = parsing_configs, date_config_list = date_configs, field_config_list = field_configs, agg_fun = fb_summarise, group_vec = "ad_set_name")
}

