% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/start_of_month.R
\name{start_of_month}
\alias{start_of_month}
\title{First date of calendar month}
\usage{
start_of_month(date)
}
\arguments{
\item{date}{A 'Date' object, or character string in the format "\%Y-\%m-\%d" or "\%Y/\%m/\%d"}
}
\value{
A Date object that is the first date of the respective calendar month.
}
\description{
First date of calendar month
}
\examples{
start_of_month("2014/02/14")
}

