% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/strip_creative_size.R
\name{strip_creative_size}
\alias{strip_creative_size}
\title{Display strip creative size}
\usage{
strip_creative_size(colname)
}
\arguments{
\item{colname}{field whose elements should be stripped of creative size}
}
\value{
a vector stripped of creative size
}
\description{
Given an input vector, strips common elements designating creative sizes for display
}

