% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summarise_dfa.R
\name{summarise_dfa}
\alias{summarise_dfa}
\title{Calculate common display aggregates on a DoubleClick data frame}
\usage{
summarise_dfa(df)
}
\arguments{
\item{df}{data frame with DoubleClick report data}
}
\value{
summarised tbl_df
}
\description{
Calculate common display aggregates on a DoubleClick data frame
}

