% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summarise_kenshoo_metrics.R
\name{summarise_kenshoo_metrics}
\alias{summarise_kenshoo_metrics}
\title{Summarise Kenshoo Metrics}
\usage{
summarise_kenshoo_metrics(df)
}
\arguments{
\item{df}{A data frame with metrics following Kenshoo conventions}
}
\value{
An aggregated data frame with metrics following Kenshoo conventions, aggregated.
}
\description{
Summarise Kenshoo Metrics
}

