% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/theme_nice.R
\name{theme_nice}
\alias{theme_nice}
\title{ggplot2 theme with pretty default settings}
\usage{
theme_nice(base_size = 20, base_family = "")
}
\arguments{
\item{base_size}{reference font size.  Default is 20.}

\item{base_family}{reference font family}
}
\value{
set of modified ggplot2 theme elements
}
\description{
ggplot2 theme with pretty default settings
}
\examples{
ggplot(mtcars, aes(x=mpg, y=wt)) + geom_point() + theme_nice(base_size=12)
}

