% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/top_n_table.R
\name{top_n_table}
\alias{top_n_table}
\title{Print a table of top n sorted variables}
\usage{
top_n_table(df, nrow = 3, pretty = TRUE, ...)
}
\arguments{
\item{df}{}

\item{nrow}{number of rows to return, after sorting}

\item{pretty}{use scales::comma on numeric values and return a pandoc.table}

\item{...}{}

\item{arrange_vec}{Character vector indicating sort order. Strings must include desc() to sort in descending order}
}
\value{
A data.frame with top n cases
}
\description{
Print a table of top n sorted variables
}
\examples{
top_n_table(mtcars, nrow = 3, pretty = FALSE)
}

